PVW-1068: Don't let the wallet and provider deal with raw hashmaps for exchanging error data.

See merge request wallet/nl-wallet-demo-app!230

+++ wallet_core/flutter_api/src/models/unlock.rs
@@ -3,7 +3,7 @@ use wallet::wallet::WalletUnlockError;
 pub enum WalletUnlockResult {
     Ok,
     IncorrectPin {
-        leftover_attempts: u64,
+        leftover_attempts: u8,
         is_final_attempt: bool,
     },
     Timeout {
